Added GSM

+++ Files/LM35/lm35.c
@@ -2,7 +2,7 @@
 #include "adc.h"
 #include "types.h"
 
-u8 ReadTempLM35(u8 chNo)
+f32 ReadTempLM35(u8 chNo)
 {
 	f32 eAR;
 	eAR=Read_ADC(chNo) * 100;
